+++ Enums.cs
@@ -1,0 +1,7 @@
+﻿namespace PeakGeneralImprovements
+{
+    public static class Enums
+    {
+        public enum eRopeVineChainOptions { Vanilla, AllowClimbing, AutoDismount }
+    }
+}