+++ Patches/SteamLobbyHandlerPatch.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using HarmonyLib;
+using PeakGeneralImprovements.Patches.Shared;
+
+namespace PeakGeneralImprovements.Patches
+{
+    internal static class SteamLobbyHandlerPatch
+    {
+        [HarmonyPatch(typeof(SteamLobbyHandler), "OnLobbyCreated")]
+        [HarmonyTranspiler]
+        private static IEnumerable<CodeInstruction> OnLobbyCreated_Transpiler(IEnumerable<CodeInstruction> instructions)
+        {
+            CodeMatcher matcher = new CodeMatcher(instructions);
+            AirportLobbySkip.ReplaceBasicLoadingScreenWithPlane(matcher, "SteamLobbyHandler.OnLobbyCreated");
+            AirportLobbySkip.TranspileStartGameButtons(matcher);
+            return matcher.InstructionEnumeration();
+        }
+    }
+}